changes after 3 lectures

+++ time_experiment.hh
@@ -1,8 +1,8 @@
 #ifndef HASC_TIME_EXPERIMENT_HH
 #define HASC_TIME_EXPERIMENT_HH
 
-#include <chrono> 
-#include <utility> 
+#include <chrono>
+#include <utility>
 
 /*! \brief time an experiment
  * This function template takes an experiment, which is a class with a method
@@ -18,42 +18,42 @@
  * \param mintime minimum total runtime (given in microseconds)
  *
  */
-template<typename T>
-auto time_experiment (const T& experiment, int mintime=250000)
+template <typename T>
+auto time_experiment(const T &experiment, int mintime = 250000)
 {
   auto start = std::chrono::high_resolution_clock::now();
   auto stop = start;
-  auto duration = stop-start;
+  auto duration = stop - start;
   auto dcast = std::chrono::duration_cast<std::chrono::microseconds>(duration).count();
-  std::pair<int,decltype(dcast)> rv;
-  int rep = 1;
-  while (dcast<mintime && rep<1000000000)
-    {
-      start = std::chrono::high_resolution_clock::now();
-      for (int i=0; i<rep; i++)
-	experiment.run();
-      stop = std::chrono::high_resolution_clock::now();
-      duration = stop-start;
-      dcast = std::chrono::duration_cast<std::chrono::microseconds>(duration).count();
-      //std::cout << " rep=" << rep << " d=" << dcast << std::endl;
-      rv.first = rep;
-      rv.second = dcast;
-      rep *= 2;
-    }
+  std::pair<long, decltype(dcast)> rv;
+  long rep = 1;
+  while (dcast < mintime && rep < 1000000000)
+  {
+    start = std::chrono::high_resolution_clock::now();
+    for (long i = 0; i < rep; i++)
+      experiment.run();
+    stop = std::chrono::high_resolution_clock::now();
+    duration = stop - start;
+    dcast = std::chrono::duration_cast<std::chrono::microseconds>(duration).count();
+    // std::cout << " rep=" << rep << " d=" << dcast << std::endl;
+    rv.first = rep;
+    rv.second = dcast;
+    rep *= 2;
+  }
   return rv;
 }
 
-auto get_time_stamp ()
+auto get_time_stamp()
 {
   return std::chrono::high_resolution_clock::now();
 }
 
-template<typename T>
-double get_duration_seconds (T start, T stop)
+template <typename T>
+double get_duration_seconds(T start, T stop)
 {
-  auto duration = stop-start;
+  auto duration = stop - start;
   auto dcast = std::chrono::duration_cast<std::chrono::microseconds>(duration).count();
-  return dcast/1e6;
+  return dcast / 1e6;
 }
 
 #endif
